Safari-611.1.10/LayoutTests/fast/css/webkit-line-clamp-calculated-value.html

<!DOCTYPE html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<div id="testDiv"></div>
<script>
description("Tests assigning a calculated value to -webkit-line-clamp CSS property.");

var testDiv = document.getElementById("testDiv");

shouldBeEmptyString("testDiv.style['-webkit-line-clamp']");
evalAndLog("testDiv.style['-webkit-line-clamp'] = 'calc(10% * 2)'");
shouldBeEqualToString("testDiv.style['-webkit-line-clamp']", "calc(20%)");
shouldBeEqualToString("window.getComputedStyle(testDiv).getPropertyValue('-webkit-line-clamp')", "20%");
evalAndLog("testDiv.style['-webkit-line-clamp'] = 'calc(2 * 3)'");
shouldBeEqualToString("testDiv.style['-webkit-line-clamp']", "6");
shouldBeEqualToString("window.getComputedStyle(testDiv).getPropertyValue('-webkit-line-clamp')", "6");

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
